+++ packages/amesos/src/Amesos_Umfpack.cpp
@@ -504,7 +504,7 @@ void Amesos_Umfpack::PrintStatus() const
   cout << "Amesos_Umfpack : Nonzero elements per row = "
        << 1.0*numentries_/NumGlobalElements_ << endl;
   cout << "Amesos_Umfpack : Percentage of nonzero elements = "
-       << 100.0*numentries_/(pow(NumGlobalElements_,2.0)) << endl;
+       << 100.0*numentries_/(pow(double(NumGlobalElements_),double(2.0))) << endl;
   cout << "Amesos_Umfpack : Use transpose = " << UseTranspose_ << endl;
 
   PrintLine();
